backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Validation & Config
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# CSV
pandas==2.1.4

# AI - Gemini
google-generativeai==0.8.5

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.25.2

# Utils
python-dateutil==2.8.2
